Refactor file to match class
