+++ src/straw/rice/__init__.py
@@ -1,1 +1,1 @@
-from .rice import Ricer
+from .ricer import Ricer
